Implemented token authentication on routes.

+++ src/controllers/app.controller.ts
@@ -2,10 +2,12 @@ import { NextFunction, Request, Response } from "express";
 import { AppService } from "../services/app.service";
 import { UserApp } from "../models/postgres/UserApp";
 import { returnCodeChallange } from "./auth.controller";
+import { AuthError } from "../errors/auth.error";
 
 
 export const appCreateController = async (req: Request, res: Response, next: NextFunction) => {
     const appData = req.body;
+    if (!req.auth) throw new AuthError("Authentication was not provided", 401);
     try {
         const appService = new AppService();
         const newApp = await appService.createApp(appData);
@@ -17,6 +19,7 @@ export const appCreateController = async (req: Request, res: Response, next: Nex
 
 export const appListController = async (req: Request, res: Response, next: NextFunction) => {
     const filter = req.query || {};
+    if (!req.auth) throw new AuthError("Authentication was not provided", 401);
     try {
         const appService = new AppService();
         const apps = await appService.getApps(filter as any);
@@ -42,6 +45,7 @@ export const appGetController = async (req: Request, res: Response, next: NextFu
 export const appUpdateController = async (req: Request, res: Response, next: NextFunction) => {
     const { id } = req.params;
     const data = req.body;
+    if (!req.auth) throw new AuthError("Authentication was not provided", 401);
     try {
         const appService = new AppService();
         const updated = await appService.updateApp(id, data);
@@ -53,6 +57,7 @@ export const appUpdateController = async (req: Request, res: Response, next: Nex
 
 export const appDeleteController = async (req: Request, res: Response, next: NextFunction) => {
     const { id } = req.params;
+    if (!req.auth) throw new AuthError("Authentication was not provided", 401);
     try {
         const appService = new AppService();
         await appService.deleteApp(id);
@@ -64,6 +69,7 @@ export const appDeleteController = async (req: Request, res: Response, next: Nex
 
 export const appCreateUserController = async (req: Request, res: Response, next: NextFunction) => {
     const { user_id, app_id } = req.body;
+    if (!req.auth) throw new AuthError("Authentication was not provided", 401);
     try {
         const appService = new AppService();
         const userApp: UserApp = await appService.createUserApp(user_id, app_id);
@@ -75,6 +81,7 @@ export const appCreateUserController = async (req: Request, res: Response, next:
 
 export const changeAppSecretController = async (req: Request, res: Response, next: NextFunction) => {
     const payload = req.body;
+    if (!req.auth) throw new AuthError("Authentication was not provided", 401);
     try {
         const appService = new AppService();
         const apps = await appService.changeAppSecret(payload);
