+++ src/controllers/device.controller.ts
@@ -1,6 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import { DeviceService } from "../services/device.service";
 import { returnCodeChallange } from "./auth.controller";
+import { AuthError } from "../errors/auth.error";
 
 export const deviceCreateController = async (req: Request, res: Response, next: NextFunction) => {
   const data = req.body;
@@ -18,6 +19,7 @@ export const deviceCreateController = async (req: Request, res: Response, next: 
 export const deviceListController = async (req: Request, res: Response, next: NextFunction) => {
   const filter = req.query as any;
   const code_challange = req.auth?.code_challenger;
+  if (!req.auth) throw new AuthError("Authentication was not provided", 401);
   try {
     
     const service = new DeviceService();
@@ -32,6 +34,7 @@ export const deviceListController = async (req: Request, res: Response, next: Ne
 
 export const deviceGetController = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
+  if (!req.auth) throw new AuthError("Authentication was not provided", 401);
   try {
     const service = new DeviceService();
     const device = await service.getDeviceById(id);
@@ -44,6 +47,7 @@ export const deviceGetController = async (req: Request, res: Response, next: Nex
 export const deviceUpdateController = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
   const data = req.body;
+  if (!req.auth) throw new AuthError("Authentication was not provided", 401);
   try {
     const service = new DeviceService();
     const updated = await service.updateDevice(id, data);
@@ -55,6 +59,7 @@ export const deviceUpdateController = async (req: Request, res: Response, next: 
 
 export const deviceDeleteController = async (req: Request, res: Response, next: NextFunction) => {
   const filter = req.query as any;
+  if (!req.auth) throw new AuthError("Authentication was not provided", 401);
   try {
     const service = new DeviceService();
     const deletedCount = await service.deleteByFilter({ device_id: filter.device_id, user_id: filter.user_id });
